Update patch set 2
Patch Set 2:

The reason I choose 18% and swappiness of 30 is that that are default values in CyanogenMod. I have been testing with 30% size, but that gave lags every now-and-then. So I went back to 18%. For swappiness, I actually prefer 60 myself, and that is the value I have been using for some time myself already.

Patch-set: 2
Label: Code-Review=0